Temporary debug: print to SDTDOUT if AssocTypePvdr.java takes more than 0.5 seconds
I have very nasty freezes several times a day, and i see the AssocTypePvdr in IDEA freeze logs. It seems that it hangs at java.io.FileInputStream.readBytes(Native Method), not sure whether it is triggered by my plugin or not. The plugin itself in the logs is shown in status WAITING while resolving some IDEA internal stuff at com.intellij.util.UniqueResultsQuery.findAll(UniqueResultsQuery.java:80). It is not the first time I have problems with this findAll(), maybe just grep for $this->{fieldName} instead of using it?